Break out rootBinding object.
